iio: adc: palmas_gpadc: use proper struct for events
palmas_gpadc can get adc event configurations from two sources:

* wakeup event data from device tree. This data is static.
* iio event data from userspace. This overwrites the device tree data.

Support for wakeup events came first and IIO event support expanded upon
this, re-using the same structures i.e. struct palmas_adc_wakeup_property.

However, this proved cumbersome. For example, to figure our if an adc
event was configured, either the low or high threshold value of that event
had to be set to a non-zero value, but a threshold value of 0 was still a
valid setting. This means that you could really confuse the code by
setting a threshold value to 0. When that happens, only 1 channel appears
to be enabled but you still cannot enable a second channel. There was also
no way to change this threshold value afterwards.

Instead of trying to fix all these corner cases in some convoluted way,
define a new struct palmas_adc_event for configuring the two events and a
helper function to convert it to struct palmas_adc_wakeup_property when
needed. This keeps the backwards compatibility with wakeup data events.
It also simplifies the code quite a bit.

Signed-off-by: Patrik Dahlström <risca@dalakolonin.se>